Split comptime intrinsic gate tests
